+++ code/backend/batch/utilities/helpers/SharePointHelper.py
